move tests to dedicated folder
